+++ NINA/ViewModel/AutoFocusVM.cs
@@ -17,6 +17,14 @@
 
     You should have received a copy of the GNU General Public License
     along with N.I.N.A..  If not, see <http://www.gnu.org/licenses/>.
+
+    Hyperbolic fitting based on CCDCiel source, also under GPL3
+    Copyright (C) 2018 Patrick Chevalley & Han Kleijn (author)
+
+    http://www.ap-i.net
+    h@ap-i.net
+
+    http://www.hnsky.org
 */
 
 #endregion "copyright"
@@ -30,6 +38,8 @@ using NINA.Utility.Mediator.Interfaces;
 using NINA.Utility.Notification;
 using NINA.Profile;
 using OxyPlot;
+using OxyPlot.Series;
+using Accord.Statistics.Models.Regression.Linear;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -38,6 +48,7 @@ using System.Threading.Tasks;
 using System.Windows.Input;
 using NINA.Utility.ImageAnalysis;
 using NINA.Model.ImageData;
+using Accord.Statistics.Models.Regression.Fitting;
 
 namespace NINA.ViewModel {
 
@@ -79,8 +90,8 @@ namespace NINA.ViewModel {
             this.guiderMediator = guiderMediator;
             this.applicationStatusMediator = applicationStatusMediator;
 
-            FocusPoints = new AsyncObservableCollection<DataPoint>();
-            
+            FocusPoints = new AsyncObservableCollection<ScatterErrorPoint>();
+            PlotFocusPoints = new AsyncObservableCollection<DataPoint>();
 
             StartAutoFocusCommand = new AsyncCommand<bool>(
                 () =>
@@ -97,7 +108,7 @@ namespace NINA.ViewModel {
                 () =>
                     Task.Run(
                         async () => {
-                            return await StartBacklashMeasurement(CommandInitializization(),_autoFocusCancelToken.Token, new Progress<ApplicationStatus>(p => Status = p));
+                            return await StartBacklashMeasurement(CommandInitializization(), _autoFocusCancelToken.Token, new Progress<ApplicationStatus>(p => Status = p));
                         }
                     ),
                 (p) => { return focuserInfo?.Connected == true && cameraInfo?.Connected == true; }
@@ -106,14 +117,15 @@ namespace NINA.ViewModel {
         }
 
         private CancellationTokenSource _autoFocusCancelToken;
-        private AsyncObservableCollection<DataPoint> _focusPoints;
+        private AsyncObservableCollection<ScatterErrorPoint> _focusPoints;
+        private AsyncObservableCollection<DataPoint> _plotFocusPoints;
         private ICameraMediator cameraMediator;
         private IImagingMediator imagingMediator;
         private IGuiderMediator guiderMediator;
         private IApplicationStatusMediator applicationStatusMediator;
-        private List<AForge.Point> brightestStarPositions = new List<AForge.Point>();
+        private List<Accord.Point> brightestStarPositions = new List<Accord.Point>();
 
-        public AsyncObservableCollection<DataPoint> FocusPoints {
+        public AsyncObservableCollection<ScatterErrorPoint> FocusPoints {
             get {
                 return _focusPoints;
             }
@@ -123,7 +135,17 @@ namespace NINA.ViewModel {
             }
         }
 
-        private DataPoint _minimum;
+        public AsyncObservableCollection<DataPoint> PlotFocusPoints {
+            get {
+                return _plotFocusPoints;
+            }
+            set {
+                _plotFocusPoints = value;
+                RaisePropertyChanged();
+            }
+        }
+
+        private ScatterErrorPoint _minimum;
 
         private ApplicationStatus _status;
 
@@ -164,6 +186,78 @@ namespace NINA.ViewModel {
             }
         }
 
+        private Func<double, double> _quadraticFitting;
+
+        public Func<double, double> QuadraticFitting {
+            get {
+                return _quadraticFitting;
+            }
+            set {
+                _quadraticFitting = value;
+                RaisePropertyChanged();
+            }
+        }
+
+        private DataPoint _quadraticMinimum;
+
+        public DataPoint QuadraticMinimum {
+            get {
+                return _quadraticMinimum;
+            }
+            set {
+                _quadraticMinimum = value;
+                RaisePropertyChanged();
+            }
+        }
+
+        private Func<double, double> _hyperbolicFitting;
+
+        public Func<double, double> HyperbolicFitting {
+            get {
+                return _hyperbolicFitting;
+            }
+            set {
+                _hyperbolicFitting = value;
+                RaisePropertyChanged();
+            }
+        }
+
+        private DataPoint _hyperbolicMinimum;
+
+        public DataPoint HyperbolicMinimum {
+            get {
+                return _hyperbolicMinimum;
+            }
+            set {
+                _hyperbolicMinimum = value;
+                RaisePropertyChanged();
+            }
+        }
+
+        private DataPoint _trendLineIntersection;
+
+        public DataPoint TrendLineIntersection {
+            get {
+                return _trendLineIntersection;
+            }
+            set {
+                _trendLineIntersection = value;
+                RaisePropertyChanged();
+            }
+        }
+
+        private DataPoint _finalFocusPoint;
+
+        public DataPoint FinalFocusPoint {
+            get {
+                return _finalFocusPoint;
+            }
+            set {
+                _finalFocusPoint = value;
+                RaisePropertyChanged();
+            }
+        }
+
         private int _focusPosition;
 
         private FilterInfo CommandInitializization() {
@@ -185,15 +279,17 @@ namespace NINA.ViewModel {
             }
 
             var comparer = new FocusPointComparer();
+            var plotComparer = new PlotPointComparer();
 
             for (int i = 0; i < nrOfSteps; i++) {
                 token.ThrowIfCancellationRequested();
 
-                double hfr = await GetAverageHFR(filter, profileService.ActiveProfile.FocuserSettings.AutoFocusNumberOfFramesPerPoint, token, progress);
+                HFRAndError measurement = await GetAverageMeasurement(filter, profileService.ActiveProfile.FocuserSettings.AutoFocusNumberOfFramesPerPoint, token, progress);
 
                 token.ThrowIfCancellationRequested();
 
-                FocusPoints.AddSorted(new DataPoint(_focusPosition, hfr), comparer);
+                FocusPoints.AddSorted(new ScatterErrorPoint(_focusPosition, measurement.HFR, 0, Math.Max(0.001, measurement.Stdev)), comparer);
+                PlotFocusPoints.AddSorted(new DataPoint(_focusPosition, measurement.HFR), plotComparer);
                 if (i < nrOfSteps - 1) {
                     Logger.Trace("Moving focuser to next autofocus position");
                     _focusPosition = await focuserMediator.MoveFocuserRelative(-stepSize);
@@ -201,6 +297,8 @@ namespace NINA.ViewModel {
 
                 token.ThrowIfCancellationRequested();
                 CalculateTrends();
+                if (FocusPoints.Count() >= 3 && (profileService.ActiveProfile.FocuserSettings.AutoFocusCurveFitting == Utility.Enum.AFCurveFittingEnum.PARABOLIC || profileService.ActiveProfile.FocuserSettings.AutoFocusCurveFitting == Utility.Enum.AFCurveFittingEnum.TRENDPARABOLIC)) { CalculateQuadraticFitting(); }
+                if (FocusPoints.Count() >= 3 && (profileService.ActiveProfile.FocuserSettings.AutoFocusCurveFitting == Utility.Enum.AFCurveFittingEnum.HYPERBOLIC || profileService.ActiveProfile.FocuserSettings.AutoFocusCurveFitting == Utility.Enum.AFCurveFittingEnum.TRENDHYPERBOLIC)) { CalculateHyperbolicFitting(); }
             }
         }
 
@@ -217,7 +315,7 @@ namespace NINA.ViewModel {
             var oldAutoStretch = imagingMediator.SetAutoStretch(true);
             var oldDetectStars = imagingMediator.SetDetectStars(false);
             IImageData image;
-            try { 
+            try {
                 image = await imagingMediator.CaptureAndPrepareImage(seq, token, progress);
             } catch (Exception e) {
                 Logger.Warning("Camera error, trying without subsample");
@@ -232,7 +330,7 @@ namespace NINA.ViewModel {
             return image;
         }
 
-        private async Task<double> EvaluateExposure(IImageData image, CancellationToken token, IProgress<ApplicationStatus> progress) {
+        private async Task<HFRAndError> EvaluateExposure(IImageData image, CancellationToken token, IProgress<ApplicationStatus> progress) {
             Logger.Trace("Evaluating Exposure");
             System.Windows.Media.PixelFormat pixelFormat;
 
@@ -265,13 +363,13 @@ namespace NINA.ViewModel {
 
             Logger.Debug(string.Format("Current Focus: Position: {0}, HRF: {1}", _focusPosition, analysis.AverageHFR));
 
-            return analysis.AverageHFR;
+            return new HFRAndError() { HFR = analysis.AverageHFR, Stdev = analysis.HFRStdDev };
         }
 
         private async Task<bool> ValidateCalculatedFocusPosition(DataPoint focusPoint, FilterInfo filter, CancellationToken token, IProgress<ApplicationStatus> progress, double initialHFR) {
             _focusPosition = await focuserMediator.MoveFocuser((int)focusPoint.X);
 
-            double hfr = await GetAverageHFR(filter, profileService.ActiveProfile.FocuserSettings.AutoFocusNumberOfFramesPerPoint, token, progress);
+            double hfr = (await GetAverageMeasurement(filter, profileService.ActiveProfile.FocuserSettings.AutoFocusNumberOfFramesPerPoint, token, progress)).HFR;
 
             if (hfr > (focusPoint.Y * 1.25)) {
                 Notification.ShowWarning(string.Format(Locale.Loc.Instance["LblFocusPointValidationFailed"], focusPoint.X, focusPoint.Y, hfr));
@@ -285,26 +383,129 @@ namespace NINA.ViewModel {
             return true;
         }
 
+        /// <summary>
+        /// Calculate HFR from position and perfectfocusposition using hyperbola parameters
+        /// The HFR of the imaged star disk as function of the focuser position can be described as hyperbola
+        /// A hyperbola is defined as:
+        /// x=b*sinh(t)
+        /// y=a*cosh(t)
+        /// Using the arccosh and arsinh functions it is possible to inverse
+        /// above calculations and convert x=>t and t->y or y->t and t->x
+        /// </summary>
+        /// <param name="position">Current focuser position</param>
+        /// <param name="perfectFocusPosition">Focuser position where HFR is lowest</param>
+        /// <param name="a">Hyperbola parameter a, lowest HFR value at focus position</param>
+        /// <param name="b">Hyperbola parameter b, defining the asymptotes, y = +-x*a/b</param>
+        /// <returns></returns>
+        private double HfrCalc(double position, double perfectFocusPosition, double a, double b) {
+            double x = perfectFocusPosition - position;
+            double t = MathHelper.HArcsin(x / b); //calculate t-position in hyperbola
+            return a * MathHelper.HCos(t); //convert t-position to y/hfd value
+        }
+
+        private double ScaledErrorHyperbola(double perfectFocusPosition, double a, double b) {
+            return Math.Sqrt(FocusPoints.Sum((dp) => Math.Pow((HfrCalc(dp.X, perfectFocusPosition, a, b) - dp.Y) / dp.ErrorY, 2)));
+        }
+
+        /// <summary>
+        /// The routine will try to find the best hyperbola curve fit. The focuser position p at the hyperbola minimum is the expected best focuser position
+        /// The FocusPoints List will be used as input to the fitting
+        /// </summary>
+        private void CalculateHyperbolicFitting() {
+            double error1, oldError, pRange, aRange, bRange, highestHfr, lowestHfr, highestPosition, lowestPosition, a, b, p, a1, b1, p1, a0, b0, p0;
+            double lowestError = double.MaxValue; //scaled RMS (square root of the mean square) of the HFD errors after curve fitting
+            int n = FocusPoints.Count();
+            ScatterErrorPoint lowestPoint = FocusPoints.Where((dp) => dp.Y >= 0.1).Aggregate((l, r) => l.Y < r.Y ? l : r); // Get lowest non-zero datapoint
+            ScatterErrorPoint highestPoint = FocusPoints.Aggregate((l, r) => l.Y > r.Y ? l : r); // Get highest datapoint
+            highestPosition = highestPoint.X;
+            highestHfr = highestPoint.Y;
+            lowestPosition = lowestPoint.X;
+            lowestHfr = lowestPoint.Y;
+            oldError = double.MaxValue;
+
+            if (highestPosition < lowestPosition) { highestPosition = 2 * lowestPosition - highestPosition; } // Always go up
+
+            //get good starting values for a, b and p
+            a = lowestHfr; // a is near the lowest HFR value
+            //Alternative hyperbola formula: sqr(y)/sqr(a)-sqr(x)/sqr(b)=1 ==>  sqr(b)=sqr(x)*sqr(a)/(sqr(y)-sqr(a)
+            b = Math.Sqrt((highestPosition - lowestPosition) * (highestPosition - lowestPosition) * a * a / (highestHfr * highestHfr - a * a));
+            p = lowestPosition;
+
+            int iterationCycles = 0; //how many cycles where used for curve fitting
+
+            //set starting test range
+            aRange = a;
+            bRange = b;
+            pRange = highestPosition - lowestPosition; //large steps since slope could contain some error
+
+            do {
+                p0 = p;
+                b0 = b;
+                a0 = a;
+
+                //Reduce range by 50%
+                aRange = aRange * 0.5;
+                bRange = bRange * 0.5;
+                pRange = pRange * 0.5;
+
+                p1 = p0 - pRange; //Start value
+
+                while (p1 <= p0 + pRange) { //Position loop
+                    a1 = a0 - aRange; //Start value
+                    while (a1 <= a0 + aRange) { //a loop
+                        b1 = b0 - bRange; // Start value
+                        while (b1 <= b0 + bRange) { //b loop
+                            error1 = ScaledErrorHyperbola(p1, a1, b1);
+                            if (error1 < lowestError) { //Better position found
+                                oldError = lowestError;
+                                lowestError = error1;
+                                //Best value up to now
+                                a = a1;
+                                b = b1;
+                                p = p1;
+                            }
+                            b1 = b1 + bRange * 0.1; //do 20 steps within range, many steps guarantees convergence
+                        }
+                        a1 = a1 + aRange * 0.1; //do 20 steps within range
+                    }
+                    p1 = p1 + pRange * 0.1; //do 20 steps within range
+                }
+                iterationCycles++;
+            } while (oldError - lowestError >= 0.0001 && lowestError > 0.0001 && iterationCycles < 30);
+            HyperbolicFitting = (x) => a * MathHelper.HCos(MathHelper.HArcsin((p - x) / b));
+            HyperbolicMinimum = new DataPoint((int)Math.Round(p), a);
+        }
+
+        private void CalculateQuadraticFitting() {
+            var fitting = new PolynomialLeastSquares() { Degree = 2 };
+            PolynomialRegression poly = fitting.Learn(FocusPoints.Select((dp) => dp.X).ToArray(), FocusPoints.Select((dp) => dp.Y).ToArray(), FocusPoints.Select((dp) => 1 / (dp.ErrorY * dp.ErrorY)).ToArray());
+            QuadraticFitting = (x) => (poly.Weights[0] * x * x + poly.Weights[1] * x + poly.Intercept);
+            int minimumX = (int)Math.Round(poly.Weights[1] / (2 * poly.Weights[0]) * -1);
+            double minimumY = QuadraticFitting(minimumX);
+            QuadraticMinimum = new DataPoint(minimumX, minimumY);
+        }
+
         private void CalculateTrends() {
             _minimum = FocusPoints.Aggregate((l, r) => l.Y < r.Y ? l : r);
-            IEnumerable<DataPoint> leftTrendPoints = FocusPoints.Where((x) => x.X < _minimum.X && x.Y > (_minimum.Y + 0.1));
-            IEnumerable<DataPoint> rightTrendPoints = FocusPoints.Where((x) => x.X > _minimum.X && x.Y > (_minimum.Y + 0.1));
+            IEnumerable<ScatterErrorPoint> leftTrendPoints = FocusPoints.Where((x) => x.X < _minimum.X && x.Y > (_minimum.Y + 0.1));
+            IEnumerable<ScatterErrorPoint> rightTrendPoints = FocusPoints.Where((x) => x.X > _minimum.X && x.Y > (_minimum.Y + 0.1));
             LeftTrend = new TrendLine(leftTrendPoints);
             RightTrend = new TrendLine(rightTrendPoints);
         }
 
-        private async Task<double> GetAverageHFR(FilterInfo filter, int exposuresPerFocusPoint, CancellationToken token, IProgress<ApplicationStatus> progress) {
-
+        private async Task<HFRAndError> GetAverageMeasurement(FilterInfo filter, int exposuresPerFocusPoint, CancellationToken token, IProgress<ApplicationStatus> progress) {
             //Average HFR  of multiple exposures (if configured this way)
             double sumHfr = 0;
+            double sumVariances = 0;
             for (int i = 0; i < exposuresPerFocusPoint; i++) {
                 var image = await TakeExposure(filter, token, progress);
-                var partialHfr = await EvaluateExposure(image, token, progress);
-                sumHfr = sumHfr + partialHfr;
+                var partialMeasurement = await EvaluateExposure(image, token, progress);
+                sumHfr = sumHfr + partialMeasurement.HFR;
+                sumVariances = sumVariances + partialMeasurement.Stdev * partialMeasurement.Stdev;
                 token.ThrowIfCancellationRequested();
             }
 
-            return sumHfr / exposuresPerFocusPoint;
+            return new HFRAndError() { HFR = sumHfr / exposuresPerFocusPoint, Stdev = Math.Sqrt(sumVariances / exposuresPerFocusPoint) };
         }
 
         public enum Direction {
@@ -322,7 +523,7 @@ namespace NINA.ViewModel {
             if (startBacklashDiag == System.Windows.MessageBoxResult.Cancel) {
                 return false;
             }
-            
+
             //Save previous backlash values
             int oldBacklashIn = profileService.ActiveProfile.FocuserSettings.BacklashIn;
             int oldBacklashOut = profileService.ActiveProfile.FocuserSettings.BacklashOut;
@@ -348,9 +549,10 @@ namespace NINA.ViewModel {
                     profileService.ActiveProfile.FocuserSettings.BacklashIn = oldBacklashIn;
                     profileService.ActiveProfile.FocuserSettings.BacklashOut = oldBacklashOut;
                 }
-            } catch(OperationCanceledException) {
+            } catch (OperationCanceledException) {
                 FocusPoints.Clear();
-            } catch(Exception e) {
+                PlotFocusPoints.Clear();
+            } catch (Exception e) {
                 Logger.Warning(e.Message);
                 Notification.ShowError(Locale.Loc.Instance["LblBacklashMeasurementException"]);
             } finally {
@@ -362,6 +564,7 @@ namespace NINA.ViewModel {
 
         public async Task<int> MeasureBacklash(FilterInfo filter, Direction direction, CancellationToken token, IProgress<ApplicationStatus> progress) {
             FocusPoints.Clear();
+            PlotFocusPoints.Clear();
             int stepSize = profileService.ActiveProfile.FocuserSettings.AutoFocusStepSize;
             int offset = profileService.ActiveProfile.FocuserSettings.AutoFocusInitialOffsetSteps;
             int backlash = 0;
@@ -370,35 +573,34 @@ namespace NINA.ViewModel {
             _focusPosition = await focuserMediator.MoveFocuserRelative((int)Math.Ceiling(offset * stepSize * 2d * (int)direction));
             token.ThrowIfCancellationRequested();
             //get HFR at this point
-            double hfr0 = await GetAverageHFR(filter, 3, token, progress);
+            double hfr0 = (await GetAverageMeasurement(filter, 3, token, progress)).HFR;
             token.ThrowIfCancellationRequested();
-            FocusPoints.AddSorted(new DataPoint(_focusPosition, hfr0), comparer);
+            FocusPoints.AddSorted(new ScatterErrorPoint(_focusPosition, hfr0, 0, 0), comparer);
 
             int counter = 0;
             double hfr1 = 0;
-            do { 
+            do {
                 //Move back one step
                 _focusPosition = await focuserMediator.MoveFocuserRelative((int)Math.Round(stepSize * (int)direction * -1d));
                 token.ThrowIfCancellationRequested();
                 //get HFR at this point
-                hfr1 = await GetAverageHFR(filter, 3, token, progress);
+                hfr1 = (await GetAverageMeasurement(filter, 3, token, progress)).HFR;
                 token.ThrowIfCancellationRequested();
-                FocusPoints.AddSorted(new DataPoint(_focusPosition, hfr1), comparer);
+                FocusPoints.AddSorted(new ScatterErrorPoint(_focusPosition, hfr1, 0, 0), comparer);
                 counter++;
-            } while (Math.Abs((hfr0 - hfr1)/hfr1) < 0.03 && counter < 3); //Slope is almost zero, backlash not cleared yet
+            } while (Math.Abs((hfr0 - hfr1) / hfr1) < 0.03 && counter < 3); //Slope is almost zero, backlash not cleared yet
 
             //Move back one more step
             _focusPosition = await focuserMediator.MoveFocuserRelative((int)Math.Round(stepSize * (int)direction * -1d));
             token.ThrowIfCancellationRequested();
             //get HFR at this point
-            double hfr2 = await GetAverageHFR(filter, 3, token, progress);
+            double hfr2 = (await GetAverageMeasurement(filter, 3, token, progress)).HFR;
             token.ThrowIfCancellationRequested();
-            FocusPoints.AddSorted(new DataPoint(_focusPosition, hfr2), comparer);
+            FocusPoints.AddSorted(new ScatterErrorPoint(_focusPosition, hfr2, 0, 0), comparer);
 
             //This far from focus, hfr0, hfr1, and hfr2 should be on a line, let's get the slopes
             double measuredSlope = Math.Abs((hfr0 - hfr1) / (stepSize * counter));
             double idealSlope = Math.Abs((hfr1 - hfr2) / stepSize);
-
 
             if (hfr1 != hfr2 && measuredSlope < idealSlope) {
                 backlash = (int)Math.Round((1 - measuredSlope / idealSlope) * stepSize * counter);
@@ -410,14 +612,21 @@ namespace NINA.ViewModel {
         public async Task<bool> StartAutoFocus(FilterInfo filter, CancellationToken token, IProgress<ApplicationStatus> progress) {
             Logger.Trace("Starting Autofocus");
             FocusPoints.Clear();
+            PlotFocusPoints.Clear();
             LeftTrend = null;
             RightTrend = null;
-            _minimum = new DataPoint(0, 0);
+            _minimum = new ScatterErrorPoint(0, 0, 0, 0);
+            TrendLineIntersection = new DataPoint(0, 0);
+            QuadraticFitting = null;
+            QuadraticMinimum = new DataPoint(0, 0);
+            HyperbolicFitting = null;
+            HyperbolicMinimum = new DataPoint(0, 0);
+            FinalFocusPoint = new DataPoint(0, 0);
             int numberOfAttempts = 0;
             int initialFocusPosition;
             double initialHFR = 0;
 
-            System.Drawing.Rectangle oldSubSample = new System.Drawing.Rectangle(); 
+            System.Drawing.Rectangle oldSubSample = new System.Drawing.Rectangle();
 
             if (profileService.ActiveProfile.FocuserSettings.AutoFocusInnerCropRatio < 1 && profileService.ActiveProfile.FocuserSettings.AutoFocusOuterCropRatio == 1 && cameraInfo.CanSubSample) {
                 oldSubSample = new System.Drawing.Rectangle(cameraInfo.SubSampleX, cameraInfo.SubSampleY, cameraInfo.SubSampleWidth, cameraInfo.SubSampleHeight);
@@ -425,10 +634,10 @@ namespace NINA.ViewModel {
                 int subSampleHeight = (int)Math.Round(cameraInfo.YSize * profileService.ActiveProfile.FocuserSettings.AutoFocusInnerCropRatio);
                 int subSampleX = (int)Math.Round((cameraInfo.XSize - subSampleWidth) / 2.0d);
                 int subSampleY = (int)Math.Round((cameraInfo.YSize - subSampleHeight) / 2.0d);
-                try { 
+                try {
                     cameraMediator.SetSubSampleArea(subSampleX, subSampleY, subSampleWidth, subSampleHeight);
                 } catch (Exception e) {
-                    Logger.Warning("Could not set subsample of rectangle X = "+ subSampleX+", Y = "+ subSampleY+", Width = "+subSampleWidth+", Height = "+subSampleHeight);
+                    Logger.Warning("Could not set subsample of rectangle X = " + subSampleX + ", Y = " + subSampleY + ", Width = " + subSampleWidth + ", Height = " + subSampleHeight);
                     Logger.Warning(e.Message);
                     _setSubSample = false;
                 }
@@ -436,12 +645,12 @@ namespace NINA.ViewModel {
             }
 
             try {
-                if (profileService.ActiveProfile.FocuserSettings.AutoFocusDisableGuiding) { 
+                if (profileService.ActiveProfile.FocuserSettings.AutoFocusDisableGuiding) {
                     await this.guiderMediator.StopGuiding(token);
                 }
 
                 //Get initial position information, as average of multiple exposures, if configured this way
-                initialHFR = await GetAverageHFR(filter, profileService.ActiveProfile.FocuserSettings.AutoFocusNumberOfFramesPerPoint, token, progress);
+                initialHFR = (await GetAverageMeasurement(filter, profileService.ActiveProfile.FocuserSettings.AutoFocusNumberOfFramesPerPoint, token, progress)).HFR;
                 initialFocusPosition = focuserInfo.Position;
 
                 bool reattempt = false;
@@ -496,13 +705,35 @@ namespace NINA.ViewModel {
                     token.ThrowIfCancellationRequested();
 
                     //Get Trendline Intersection
-                    var p = LeftTrend.Intersect(RightTrend);
+                    TrendLineIntersection = LeftTrend.Intersect(RightTrend);
 
-                    LastAutoFocusPoint = new AutoFocusPoint { Focuspoint = p, Temperature = focuserInfo.Temperature, Timestamp = DateTime.Now };
+                    if (profileService.ActiveProfile.FocuserSettings.AutoFocusCurveFitting == Utility.Enum.AFCurveFittingEnum.TRENDLINES) {
+                        FinalFocusPoint = TrendLineIntersection;
+                    }
 
-                    //Todo when data is too noisy for trend lines find something else
+                    if (profileService.ActiveProfile.FocuserSettings.AutoFocusCurveFitting == Utility.Enum.AFCurveFittingEnum.HYPERBOLIC) {
+                        CalculateHyperbolicFitting();
+                        FinalFocusPoint = HyperbolicMinimum;
+                    }
 
-                    bool goodAutoFocus = await ValidateCalculatedFocusPosition(p, filter, token, progress, initialHFR);
+                    if (profileService.ActiveProfile.FocuserSettings.AutoFocusCurveFitting == Utility.Enum.AFCurveFittingEnum.PARABOLIC) {
+                        CalculateQuadraticFitting();
+                        FinalFocusPoint = QuadraticMinimum;
+                    }
+
+                    if (profileService.ActiveProfile.FocuserSettings.AutoFocusCurveFitting == Utility.Enum.AFCurveFittingEnum.TRENDPARABOLIC) {
+                        CalculateQuadraticFitting();
+                        FinalFocusPoint = new DataPoint(Math.Round((TrendLineIntersection.X + QuadraticMinimum.X) / 2), (TrendLineIntersection.Y + QuadraticMinimum.Y) / 2);
+                    }
+
+                    if (profileService.ActiveProfile.FocuserSettings.AutoFocusCurveFitting == Utility.Enum.AFCurveFittingEnum.TRENDHYPERBOLIC) {
+                        CalculateHyperbolicFitting();
+                        FinalFocusPoint = new DataPoint(Math.Round((TrendLineIntersection.X + HyperbolicMinimum.X) / 2), (TrendLineIntersection.Y + HyperbolicMinimum.Y) / 2);
+                    }
+
+                    LastAutoFocusPoint = new AutoFocusPoint { Focuspoint = FinalFocusPoint, Temperature = focuserInfo.Temperature, Timestamp = DateTime.Now };
+
+                    bool goodAutoFocus = await ValidateCalculatedFocusPosition(FinalFocusPoint, filter, token, progress, initialHFR);
 
                     if (!goodAutoFocus) {
                         if (numberOfAttempts < profileService.ActiveProfile.FocuserSettings.AutoFocusTotalNumberOfAttempts) {
@@ -510,9 +741,16 @@ namespace NINA.ViewModel {
                             await focuserMediator.MoveFocuser(initialFocusPosition);
                             Logger.Warning("Potentially bad auto-focus. Reattempting.");
                             FocusPoints.Clear();
+                            PlotFocusPoints.Clear();
                             LeftTrend = null;
                             RightTrend = null;
-                            _minimum = new DataPoint(0, 0);
+                            _minimum = new ScatterErrorPoint(0, 0, 0, 0);
+                            TrendLineIntersection = new DataPoint(0, 0);
+                            QuadraticFitting = null;
+                            QuadraticMinimum = new DataPoint(0, 0);
+                            HyperbolicFitting = null;
+                            HyperbolicMinimum = new DataPoint(0, 0);
+                            FinalFocusPoint = new DataPoint(0, 0);
                             reattempt = true;
                         } else {
                             Notification.ShowWarning(Locale.Loc.Instance["LblAutoFocusRestoringOriginalPosition"]);
@@ -526,15 +764,16 @@ namespace NINA.ViewModel {
                 //_focusPosition = await Mediator.Instance.RequestAsync(new MoveFocuserMessage() { Position = (int)p.X, Absolute = true, Token = token });
             } catch (OperationCanceledException) {
                 FocusPoints.Clear();
+                PlotFocusPoints.Clear();
             } catch (Exception ex) {
                 Notification.ShowError(ex.Message);
                 Logger.Error(ex);
             } finally {
                 //Restore original sub-sample rectangle, if appropriate
                 if (_setSubSample && oldSubSample.X >= 0 && oldSubSample.Y >= 0 && oldSubSample.Width > 0 && oldSubSample.Height > 0) {
-                    try { 
+                    try {
                         cameraMediator.SetSubSampleArea((int)oldSubSample.X, (int)oldSubSample.Y, (int)oldSubSample.Width, (int)oldSubSample.Height);
-                    } catch(Exception e) {
+                    } catch (Exception e) {
                         Logger.Warning("Could not set back old sub sample area");
                         Logger.Warning(e.Message);
                         Notification.ShowError(e.Message);
@@ -593,13 +832,59 @@ namespace NINA.ViewModel {
         public ICommand CancelBacklashMeasurementCommand { get; private set; }
     }
 
+    public static class MathHelper {
+
+        // Hyperbolic Sine
+        public static double HSin(double x) {
+            return (Math.Exp(x) - Math.Exp(-x)) / 2;
+        }
+
+        // Hyperbolic Cosine
+        public static double HCos(double x) {
+            return (Math.Exp(x) + Math.Exp(-x)) / 2;
+        }
+
+        // Hyperbolic Tangent
+        public static double HTan(double x) {
+            return (Math.Exp(x) - Math.Exp(-x)) / (Math.Exp(x) + Math.Exp(-x));
+        }
+
+        // Inverse Hyperbolic Sine
+        public static double HArcsin(double x) {
+            return Math.Log(x + Math.Sqrt(x * x + 1));
+        }
+
+        // Inverse Hyperbolic Cosine
+        public static double HArccos(double x) {
+            return Math.Log(x + Math.Sqrt(x * x - 1));
+        }
+
+        // Inverse Hyperbolic Tangent
+        public static double HArctan(double x) {
+            return Math.Log((1 + x) / (1 - x)) / 2;
+        }
+    }
+
     public class AutoFocusPoint {
         public DataPoint Focuspoint { get; set; }
         public DateTime Timestamp { get; set; }
         public double Temperature { get; set; }
     }
 
-    public class FocusPointComparer : IComparer<DataPoint> {
+    public class FocusPointComparer : IComparer<ScatterErrorPoint> {
+
+        public int Compare(ScatterErrorPoint x, ScatterErrorPoint y) {
+            if (x.X < y.X) {
+                return -1;
+            } else if (x.X > y.X) {
+                return 1;
+            } else {
+                return 0;
+            }
+        }
+    }
+
+    public class PlotPointComparer : IComparer<DataPoint> {
 
         public int Compare(DataPoint x, DataPoint y) {
             if (x.X < y.X) {
@@ -612,33 +897,33 @@ namespace NINA.ViewModel {
         }
     }
 
+    public struct HFRAndError {
+        public double HFR { get; set; }
+        public double Stdev { get; set; }
+    }
+
     public class TrendLine {
 
-        public TrendLine(IEnumerable<DataPoint> l) {
+        public TrendLine(IEnumerable<ScatterErrorPoint> l) {
             DataPoints = l;
 
-            var n = DataPoints.Count();
-            var sumXY = DataPoints.Sum((x) => x.X * x.Y);
-            var sumX = DataPoints.Sum((x) => x.X);
-            var sumY = DataPoints.Sum((x) => x.Y);
-            var sumXsumY = sumX * sumY;
-            var sumXsquared = DataPoints.Sum((x) => Math.Pow(x.X, 2));
+            if (DataPoints.Count() > 1) {
+                double[] inputs = DataPoints.Select((dp) => dp.X).ToArray();
+                double[] outputs = DataPoints.Select((dp) => dp.Y).ToArray();
+                double[] weights = DataPoints.Select((dp) => 1 / (dp.ErrorY * dp.ErrorY)).ToArray();
 
-            var alpha = (n * sumXY - sumXsumY) / (n * sumXsquared - Math.Pow(sumX, 2));
+                OrdinaryLeastSquares ols = new OrdinaryLeastSquares();
+                SimpleLinearRegression regression = ols.Learn(inputs, outputs, weights);
 
-            var beta = (sumY - alpha * sumX) / n;
-            var beta2 = (sumY * sumXsquared - sumX * sumXY) / (n * sumXsquared - Math.Pow(sumX, 2));
-
-            Slope = alpha;
-            Offset = beta;
-
-            // y = alpha * x + beta
+                Slope = regression.Slope;
+                Offset = regression.Intercept;
+            }
         }
 
         public double Slope { get; set; }
         public double Offset { get; set; }
 
-        public IEnumerable<DataPoint> DataPoints { get; set; }
+        public IEnumerable<ScatterErrorPoint> DataPoints { get; set; }
 
         public double GetY(double x) {
             return Slope * x + Offset;